backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langchain-openai==0.3.30
python-dotenv==1.1.1
python-pptx==1.0.2
